# swagger_server/models/shop_unit.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
#from swagger_server.models.shop_unit import ShopUnit  # noqa: F401,E501
from swagger_server.models.shop_unit_type import ShopUnitType  # noqa: F401,E501
from swagger_server import util


class ShopUnit(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, id: str=None, name: str=None, _date: str=None, parent_id: str=None, type: ShopUnitType=None, price: int=None, children: []=None):  # noqa: E501
        """ShopUnit - a model defined in Swagger

        :param id: The id of this ShopUnit.  # noqa: E501
        :type id: str
        :param name: The name of this ShopUnit.  # noqa: E501
        :type name: str
        :param _date: The _date of this ShopUnit.  # noqa: E501
        :type _date: str
        :param parent_id: The parent_id of this ShopUnit.  # noqa: E501
        :type parent_id: str
        :param type: The type of this ShopUnit.  # noqa: E501
        :type type: ShopUnitType
        :param price: The price of this ShopUnit.  # noqa: E501
        :type price: int
        :param children: The children of this ShopUnit.  # noqa: E501
        :type children: List[ShopUnit]
        """
        self.swagger_types = {
            'id': str,
            'name': str,
            '_date': str,
            'parent_id': str,
            'type': ShopUnitType,
            'price': int,
            'children': List[ShopUnit]
        }

        self.attribute_map = {
            'id': 'id',
            'name': 'name',
            '_date': 'date',
            'parent_id': 'parentId',
            'type': 'type',
            'price': 'price',
            'children': 'children'
        }
        self._id = id
        self._name = name
        self.__date = _date
        self._parent_id = parent_id
        self._type = type
        self._price = price
        self._children = children

    @classmethod
    def from_dict(cls, dikt) -> 'ShopUnit':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The ShopUnit of this ShopUnit.  # noqa: E501
        :rtype: ShopUnit
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> str:
        """Gets the id of this ShopUnit.

        Уникальный идентфикатор  # noqa: E501

        :return: The id of this ShopUnit.
        :rtype: str
        """
        return self._id

    @id.setter
    def id(self, id: str):
        """Sets the id of this ShopUnit.

        Уникальный идентфикатор  # noqa: E501

        :param id: The id of this ShopUnit.
        :type id: str
        """
        if id is None:
            raise ValueError("Invalid value for `id`, must not be `None`")  # noqa: E501

        self._id = id

    @property
    def name(self) -> str:
        """Gets the name of this ShopUnit.

        Имя категории  # noqa: E501

        :return: The name of this ShopUnit.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """Sets the name of this ShopUnit.

        Имя категории  # noqa: E501

        :param name: The name of this ShopUnit.
        :type name: str
        """
        if name is None:
            raise ValueError("Invalid value for `name`, must not be `None`")  # noqa: E501

        self._name = name

    @property
    def _date(self) -> str:
        """Gets the _date of this ShopUnit.

        Время последнего обновления элемента.  # noqa: E501

        :return: The _date of this ShopUnit.
        :rtype: str
        """
        return self.__date

    @_date.setter
    def _date(self, _date: str):
        """Sets the _date of this ShopUnit.

        Время последнего обновления элемента.  # noqa: E501

        :param _date: The _date of this ShopUnit.
        :type _date: str
        """
        if _date is None:
            raise ValueError("Invalid value for `_date`, must not be `None`")  # noqa: E501

        self.__date = _date

    @property
    def parent_id(self) -> str:
        """Gets the parent_id of this ShopUnit.

        UUID родительской категории  # noqa: E501

        :return: The parent_id of this ShopUnit.
        :rtype: str
        """
        return self._parent_id

    @parent_id.setter
    def parent_id(self, parent_id: str):
        """Sets the parent_id of this ShopUnit.

        UUID родительской категории  # noqa: E501

        :param parent_id: The parent_id of this ShopUnit.
        :type parent_id: str
        """

        self._parent_id = parent_id

    @property
    def type(self) -> ShopUnitType:
        """Gets the type of this ShopUnit.


        :return: The type of this ShopUnit.
        :rtype: ShopUnitType
        """
        return self._type

    @type.setter
    def type(self, type: ShopUnitType):
        """Sets the type of this ShopUnit.


        :param type: The type of this ShopUnit.
        :type type: ShopUnitType
        """
        if type is None:
            raise ValueError("Invalid value for `type`, must not be `None`")  # noqa: E501

        self._type = type

    @property
    def price(self) -> int:
        """Gets the price of this ShopUnit.

        Целое число, для категории - это средняя цена всех дочерних товаров(включая товары подкатегорий). Если цена является не целым числом, округляется в меньшую сторону до целого числа. Если категория не содержит товаров цена равна null.  # noqa: E501

        :return: The price of this ShopUnit.
        :rtype: int
        """
        return self._price

    @price.setter
    def price(self, price: int):
        """Sets the price of this ShopUnit.

        Целое число, для категории - это средняя цена всех дочерних товаров(включая товары подкатегорий). Если цена является не целым числом, округляется в меньшую сторону до целого числа. Если категория не содержит товаров цена равна null.  # noqa: E501

        :param price: The price of this ShopUnit.
        :type price: int
        """

        self._price = price

    @property
    def children(self) -> []:
        """Gets the children of this ShopUnit.

        Список всех дочерних товаров\\категорий. Для товаров поле равно null.  # noqa: E501

        :return: The children of this ShopUnit.
        :rtype: List[ShopUnit]
        """
        return self._children

    @children.setter
    def children(self, children: []):
        """Sets the children of this ShopUnit.

        Список всех дочерних товаров\\категорий. Для товаров поле равно null.  # noqa: E501

        :param children: The children of this ShopUnit.
        :type children: List[ShopUnit]
        """

        self._children = children